Form Updates + FPAC's Food Resources

+++ app/views.py
@@ -23,11 +23,10 @@ def index():
 @login_required
 def new(id=None):
 	form = AddNewFoodResourceForm(request.form)
-
-	# Used to dynamically pre-load the form with existing database contents.
-	timeslots = []
-	food_resource_type = ""
-	are_hours_available = False
+	for timeslots in form.daily_timeslots:
+		for timeslot in timeslots.timeslots:
+			timeslot.starts_at.choices=get_possible_opening_times()
+			timeslot.ends_at.choices=get_possible_closing_times()
 
 	# Create a new food resource. 
 	if id is None:
@@ -39,12 +38,15 @@ def new(id=None):
 
 	# GET request.
 	if request.method == 'GET' and id is not None:
+
 		# Populate form with information about existing food resource. 
 		food_resource = FoodResource.query.filter_by(id=id).first()
 		if food_resource is None:
 			return render_template('404.html')
 
+		# Data that can be directly retrieved from the database.
 		form.name.data = food_resource.name
+		form.location_type.data = food_resource.location_type
 		form.address_line1.data = food_resource.address.line1
 		form.address_line2.data = food_resource.address.line2
 		form.address_city.data = food_resource.address.city
@@ -53,41 +55,68 @@ def new(id=None):
 		form.phone_number.data = food_resource.phone_numbers[0].number
 		form.website.data = food_resource.url
 		form.additional_information.data = food_resource.description
-		form.is_for_family_and_children.data = food_resource.is_for_family_and_children
+		form.is_for_family_and_children.data = \
+			food_resource.is_for_family_and_children
 		form.is_for_seniors.data = food_resource.is_for_seniors
-		form.is_wheelchair_accessible.data = food_resource.is_wheelchair_accessible
+		form.is_wheelchair_accessible.data = \
+			food_resource.is_wheelchair_accessible
 		form.is_accepts_snap.data = food_resource.is_accepts_snap
 
-		# Fields that must be dyanamically updated using JavaScript.
+		# Data that must be interpreted before being rendered.
+		if food_resource.are_hours_available == True:
+			form.are_hours_available.data = "yes"
+		else:
+			form.are_hours_available.data = "no"
+
 		for timeslot in food_resource.timeslots:
-			timeslots.append(timeslot)
-		food_resource_type = food_resource.location_type
-		are_hours_available = food_resource.are_hours_available
+			index = timeslot.day_of_week
+			start_time = timeslot.start_time
+			end_time = timeslot.end_time
+			form.daily_timeslots[index].timeslots[0].starts_at.data = \
+				start_time.strftime("%H:%M")
+			form.daily_timeslots[index].timeslots[0].ends_at.data = \
+				end_time.strftime("%H:%M")
+			form.is_open[index].is_open.data = "open"
 
 	# POST request.
 	additional_errors = []
 	if request.method == 'POST' and form.validate(): 
-		food_resource_type = request.form['food-resource-type']
+		food_resource_type = form.location_type.data
+		all_timeslots = []
+
+		if form.are_hours_available.data == "yes":
+			are_hours_available = True
+		else:
+			are_hours_available = False
 
 		# Create the food resource's timeslots.
 		are_timeslots_valid = True
-		for i, day_of_week in enumerate(days_of_week): 
-			if (request.form[str(day_of_week['index']) + '-open-or-closed'] == "open"):
-				opening_time = request.form[str(day_of_week['index']) + '-opening-time']
-				start_time = get_time_from_string(opening_time)
-				closing_time = request.form[str(day_of_week['index']) + '-closing-time']
-				end_time = get_time_from_string(closing_time)
-				timeslot = TimeSlot(day_of_week=i, start_time=start_time, 
-					end_time=end_time)
-				timeslots.append(timeslot)
+		if are_hours_available: 
+			for i, timeslots in enumerate(form.daily_timeslots):
+				for timeslot in timeslots.timeslots:
+					# Check if food resource is open on the i-th day of the 
+					# week.
+					is_open = True
+					if form.is_open[i].is_open.data == "closed":
+						is_open = False
 
-				# Check that timeslot is valid.
-				if start_time >= end_time: 
-					are_timeslots_valid = False
-					additional_errors.append("Opening time must be before \
-						closing time.")
-				else:
-					db.session.add(timeslot)
+					# Create timeslots only if the food resource is open on the
+					# i-th day of the week.
+					if is_open:
+						start_time = \
+							get_time_from_string(timeslot.starts_at.data)
+						end_time = get_time_from_string(timeslot.ends_at.data)
+						timeslot = TimeSlot(day_of_week=i, 
+							start_time=start_time, end_time=end_time)
+						all_timeslots.append(timeslot)
+
+						# Check that timeslot is valid.
+						if start_time >= end_time: 
+							are_timeslots_valid = False
+							additional_errors.append("Opening time must be \
+								before closing time.")
+						else:
+							db.session.add(timeslot)
 
 		# Create the food resource's remaining attributes. 
 		if are_timeslots_valid:
@@ -96,9 +125,9 @@ def new(id=None):
 			# from the database. 
 			if id is not None:
 				fr = FoodResource.query.filter_by(id=id).first()
-				db.session.delete(fr)
-				db.session.commit()
-
+				if fr:
+					db.session.delete(fr)
+					db.session.commit()
 			# Create food resource's address.
 			address = Address(line1=form.address_line1.data, 
 				line2=form.address_line2.data, 
@@ -113,28 +142,20 @@ def new(id=None):
 			db.session.add(home_number)
 			phone_numbers.append(home_number)
 
-			are_hours_available = request.form['are-hours-available']
-			if are_hours_available == 'yes':
-				are_hours_available = True
-			else:
-				are_hours_available = False
-
 			# Create food resource and store all data in it.
 			food_resource = FoodResource(
 				name=form.name.data, 
 				phone_numbers=phone_numbers,
 				description=form.additional_information.data,
-				timeslots=timeslots,
+				timeslots=all_timeslots,
 				address=address, 
-				is_for_family_and_children = form.is_for_family_and_children.data,
+				is_for_family_and_children = \
+					form.is_for_family_and_children.data,
 				is_for_seniors = form.is_for_seniors.data,
 				is_wheelchair_accessible = form.is_wheelchair_accessible.data,	
 				is_accepts_snap = form.is_accepts_snap.data, 
-				are_hours_available = are_hours_available)
-
-			# Assign a type to the food resource. 
-			food_resource.location_type = request.form['food-resource-type']
-
+				are_hours_available = are_hours_available, 
+				location_type = food_resource_type)
 			# Commit all database changes. 
 			db.session.add(food_resource)
 			db.session.commit()
@@ -144,23 +165,17 @@ def new(id=None):
 	# render the page. 
 	return render_template('add_resource.html', form=form, 
 		days_of_week=days_of_week, resources_info=resources_info_singular, 
-		additional_errors=additional_errors, title=title, 
-
-		# Used to populate drop-downs for opening and closing times.
-		possible_opening_times=get_possible_opening_times(), 
-		possible_closing_times=get_possible_closing_times(),
-
-		# Used to dynamically pre-load form.
-		timeslots=timeslots, 
-		food_resource_type=food_resource_type, 
-		are_hours_available=are_hours_available)
+		additional_errors=additional_errors, title=title)
 
 #Allows non-admins to add food resources
 @app.route('/propose-resource', methods=['GET', 'POST'])
 def guest_new_food_resource():
 	form = NonAdminAddNewFoodResourceForm(request.form)
-	timeslots = []
-	food_resource_type = "FARMERS_MARKET"
+	for timeslots in form.daily_timeslots:
+		for timeslot in timeslots.timeslots:
+			timeslot.starts_at.choices=get_possible_opening_times()
+			timeslot.ends_at.choices=get_possible_closing_times()
+	form.location_type.data = "FARMERS_MARKET"
 
 	additional_errors = []
 	if request.method == 'POST' and form.validate(): 
@@ -170,37 +185,50 @@ def guest_new_food_resource():
 		guest_email = form.your_email_address.data
 		guest_phone_number = form.your_phone_number.data
 
-		#Checks to see if this contact exists
+		# Check to see if this contact exists.
 		contact = FoodResourceContact.query.filter_by(
 			email = guest_email, name = guest_name).first()
 
-		if(contact is None):
-			contact = FoodResourceContact(name = guest_name, email = guest_email,
-				phone_number = guest_phone_number)
+		if (contact is None):
+			contact = FoodResourceContact(name=guest_name, 
+				email=guest_email, phone_number=guest_phone_number)
 			db.session.add(contact)
 			db.session.commit()
 
-		food_resource_type = request.form['food-resource-type']
+		if form.are_hours_available.data == "yes":
+			are_hours_available = True
+		else:
+			are_hours_available = False
 
 		# Create the food resource's timeslots.
 		are_timeslots_valid = True
-		for i, day_of_week in enumerate(days_of_week): 
-			if (request.form[str(day_of_week['index']) + '-open-or-closed'] == "open"):
-				opening_time = request.form[str(day_of_week['index']) + '-opening-time']
-				start_time = get_time_from_string(opening_time)
-				closing_time = request.form[str(day_of_week['index']) + '-closing-time']
-				end_time = get_time_from_string(closing_time)
-				timeslot = TimeSlot(day_of_week=i, start_time=start_time, 
-					end_time=end_time)
-				timeslots.append(timeslot)
+		all_timeslots = []
+		if are_hours_available: 
+			for i, timeslots in enumerate(form.daily_timeslots):
+				for timeslot in timeslots.timeslots:
+					# Check if food resource is open on the i-th day of the 
+					# week.
+					is_open = True
+					if form.is_open[i].is_open.data == "closed":
+						is_open = False
 
-				# Check that timeslot is valid.
-				if start_time >= end_time: 
-					are_timeslots_valid = False
-					additional_errors.append("Opening time must be before \
-						closing time.")
-				else:
-					db.session.add(timeslot)
+					# Create timeslots only if the food resource is open on the
+					# i-th day of the week.
+					if is_open:
+						start_time = \
+							get_time_from_string(timeslot.starts_at.data)
+						end_time = get_time_from_string(timeslot.ends_at.data)
+						timeslot = TimeSlot(day_of_week=i, 
+							start_time=start_time, end_time=end_time)
+						all_timeslots.append(timeslot)
+
+						# Check that timeslot is valid.
+						if start_time >= end_time: 
+							are_timeslots_valid = False
+							additional_errors.append("Opening time must be \
+								before closing time.")
+						else:
+							db.session.add(timeslot)
 
 		# Create the food resource's remaining attributes. 
 		if are_timeslots_valid:
@@ -219,29 +247,22 @@ def guest_new_food_resource():
 			db.session.add(home_number)
 			phone_numbers.append(home_number)
 
-			are_hours_available = request.form['are-hours-available']
-			if are_hours_available == 'yes':
-				are_hours_available = True
-			else:
-				are_hours_available = False
-
 			# Create food resource and store all data in it.
 			food_resource = FoodResource(
-				name = form.name.data, 
-				phone_numbers = phone_numbers,
-				description = form.additional_information.data,
-				timeslots = timeslots,
-				address = address,
-				are_hours_available = are_hours_available, 
-				is_approved = False,
-				food_resource_contact = contact, 
-				is_for_family_and_children = form.is_for_family_and_children.data,
+				name=form.name.data, 
+				phone_numbers=phone_numbers,
+				description=form.additional_information.data,
+				timeslots=all_timeslots,
+				address=address, 
+				is_for_family_and_children = \
+					form.is_for_family_and_children.data,
 				is_for_seniors = form.is_for_seniors.data,
 				is_wheelchair_accessible = form.is_wheelchair_accessible.data,	
-				is_accepts_snap = form.is_accepts_snap.data)
-
-			# Assign a type to the food resource. 
-			food_resource.location_type = request.form['food-resource-type']
+				is_accepts_snap = form.is_accepts_snap.data, 
+				are_hours_available = are_hours_available, 
+				location_type = form.location_type.data, 
+				is_approved = False, 
+				food_resource_contact = contact)
 
 			# Commit all database changes. 
 			db.session.add(food_resource)
@@ -252,10 +273,7 @@ def guest_new_food_resource():
 	# render the page. 
 	return render_template('guest_add_resource.html', form=form, 
 		days_of_week=days_of_week, resources_info=resources_info_singular, 
-		additional_errors=additional_errors, timeslots=timeslots, 
-		food_resource_type=food_resource_type, 
-		possible_opening_times=get_possible_opening_times(), 
-		possible_closing_times=get_possible_closing_times())
+		additional_errors=additional_errors)
 
 @app.route('/_thank-you')
 def post_guest_add():
@@ -265,16 +283,29 @@ def post_guest_add():
 @login_required
 def admin():
 	resources = {}
-	resources['farmers-markets'] = FoodResource.query.filter_by(location_type="FARMERS_MARKET", is_approved=True).order_by(FoodResource.name).all()
-	resources['senior-meals'] = FoodResource.query.filter_by(location_type="SENIOR_MEAL", is_approved=True).order_by(FoodResource.name).all()
-	resources['food-cupboards'] = FoodResource.query.filter_by(location_type="FOOD_CUPBOARD", is_approved=True).order_by(FoodResource.name).all()
-	resources['share-host-sites'] = FoodResource.query.filter_by(location_type="SHARE", is_approved=True).order_by(FoodResource.name).all()
-	resources['soup-kitchens'] = FoodResource.query.filter_by(location_type="SOUP_KITCHEN", is_approved=True).order_by(FoodResource.name).all()
-	resources['wic-offices'] = FoodResource.query.filter_by(location_type="WIC_OFFICE", is_approved=True).order_by(FoodResource.name).all()
+	resources['farmers-markets'] = FoodResource.query \
+		.filter_by(location_type="FARMERS_MARKET", is_approved=True) \
+		.order_by(FoodResource.name).all()
+	resources['senior-meals'] = FoodResource.query \
+		.filter_by(location_type="SENIOR_MEAL", is_approved=True) \
+		.order_by(FoodResource.name).all()
+	resources['food-cupboards'] = FoodResource.query \
+		.filter_by(location_type="FOOD_CUPBOARD", is_approved=True) \
+		.order_by(FoodResource.name).all() 
+	resources['share-host-sites'] = FoodResource.query \
+		.filter_by(location_type="SHARE", is_approved=True) \
+		.order_by(FoodResource.name).all()
+	resources['soup-kitchens'] = FoodResource.query \
+		.filter_by(location_type="SOUP_KITCHEN", is_approved=True) \
+		.order_by(FoodResource.name).all()
+	resources['wic-offices'] = FoodResource.query \
+		.filter_by(location_type="WIC_OFFICE", is_approved=True) \
+		.order_by(FoodResource.name).all()
 
 	contacts = FoodResourceContact.query.all()
 
-	return render_template('admin_resources.html', food_resource_contacts = contacts,
+	return render_template('admin_resources.html', 
+		food_resource_contacts=contacts,
 		resources_info=resources_info_plural, resources=resources, 
 		days_of_week=days_of_week)
